refactor(assistant-stream): split the ui message stream chunk normalizer out of the decoder
