+++ .changeset/split-ui-message-stream-chunk-normalizer.md
@@ -1,0 +1,5 @@
+---
+"assistant-stream": patch
+---
+
+refactor: split the ui message stream chunk normalizer out of the decoder
